+++ pkg/azuredx/client/client.go
@@ -14,6 +14,7 @@ import (
 	"github.com/grafana/grafana-azure-sdk-go/v2/azsettings"
 	"github.com/grafana/grafana-azure-sdk-go/v2/azusercontext"
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
+	"github.com/grafana/grafana-plugin-sdk-go/experimental/errorsource"
 
 	// 100% compatible drop-in replacement of "encoding/json"
 	json "github.com/json-iterator/go"
@@ -198,12 +199,12 @@ func (c *Client) testManagementClient(ctx context.Context, _ *models.DatasourceS
 func (c *Client) KustoRequest(ctx context.Context, clusterUrl string, path string, payload models.RequestPayload, userTrackingEnabled bool) (*models.TableResponse, error) {
 	buf, err := json.Marshal(payload)
 	if err != nil {
-		return nil, fmt.Errorf("no Azure request serial: %w", err)
+		return nil, errorsource.DownstreamError(fmt.Errorf("no Azure request serial: %w", err), false)
 	}
 
 	fullUrl, err := url.JoinPath(clusterUrl, path)
 	if err != nil {
-		return nil, fmt.Errorf("invalid Azure request URL: %w", err)
+		return nil, errorsource.DownstreamError(fmt.Errorf("invalid Azure request URL: %w", err), false)
 	}
 
 	req, err := http.NewRequestWithContext(ctx, http.MethodPost, fullUrl, bytes.NewReader(buf))
@@ -229,7 +230,7 @@ func (c *Client) KustoRequest(ctx context.Context, clusterUrl string, path strin
 
 	resp, err := c.httpClientKusto.Do(req)
 	if err != nil {
-		return nil, err
+		return nil, errorsource.DownstreamError(err, false)
 	}
 
 	defer resp.Body.Close()
@@ -237,15 +238,15 @@ func (c *Client) KustoRequest(ctx context.Context, clusterUrl string, path strin
 	switch {
 	case resp.StatusCode == http.StatusUnauthorized:
 		// HTTP 401 has no error body
-		return nil, fmt.Errorf("azure HTTP %q", resp.Status)
+		return nil, errorsource.DownstreamError(fmt.Errorf("azure HTTP %q", resp.Status), false)
 
 	case resp.StatusCode/100 != 2:
 		var r models.ErrorResponse
 		err := json.NewDecoder(resp.Body).Decode(&r)
 		if err != nil {
-			return nil, fmt.Errorf("azure HTTP %q with malformed error response: %s", resp.Status, err)
+			return nil, errorsource.DownstreamError(fmt.Errorf("azure HTTP %q with malformed error response: %s", resp.Status, err), false)
 		}
-		return nil, fmt.Errorf("azure HTTP %q: %q.\nReceived %q: %q", resp.Status, r.Error.Message, r.Error.Type, r.Error.Description)
+		return nil, errorsource.SourceError(backend.ErrorSourceFromHTTPStatus(resp.StatusCode), fmt.Errorf("azure HTTP %q: %q.\nReceived %q: %q", resp.Status, r.Error.Message, r.Error.Type, r.Error.Description), false)
 	}
 
 	return models.TableFromJSON(resp.Body)
@@ -254,7 +255,7 @@ func (c *Client) KustoRequest(ctx context.Context, clusterUrl string, path strin
 func (c *Client) ARGClusterRequest(ctx context.Context, payload models.ARGRequestPayload, additionalHeaders map[string]string) ([]models.ClusterOption, error) {
 	buf, err := json.Marshal(payload)
 	if err != nil {
-		return nil, fmt.Errorf("no Azure request serial: %w", err)
+		return nil, errorsource.DownstreamError(fmt.Errorf("no Azure request serial: %w", err), false)
 	}
 
 	resourceManager, ok := c.cloudSettings.Properties["resourceManager"]
@@ -264,7 +265,7 @@ func (c *Client) ARGClusterRequest(ctx context.Context, payload models.ARGReques
 
 	u, err := url.Parse(resourceManager)
 	if err != nil {
-		return nil, fmt.Errorf("invalid Azure request URL: %w", err)
+		return nil, errorsource.DownstreamError(fmt.Errorf("invalid Azure request URL: %w", err), false)
 	}
 
 	// the default path for Azure Resource Graph
@@ -289,7 +290,7 @@ func (c *Client) ARGClusterRequest(ctx context.Context, payload models.ARGReques
 
 	resp, err := c.httpClientManagement.Do(req)
 	if err != nil {
-		return nil, err
+		return nil, errorsource.DownstreamError(err, false)
 	}
 
 	defer resp.Body.Close()
@@ -297,15 +298,15 @@ func (c *Client) ARGClusterRequest(ctx context.Context, payload models.ARGReques
 	switch {
 	case resp.StatusCode == http.StatusUnauthorized:
 		// HTTP 401 has no error body
-		return nil, fmt.Errorf("azure HTTP %q", resp.Status)
+		return nil, errorsource.DownstreamError(fmt.Errorf("azure HTTP %q", resp.Status), false)
 
 	case resp.StatusCode/100 != 2:
 		var r models.ErrorResponse
 		err := json.NewDecoder(resp.Body).Decode(&r)
 		if err != nil {
-			return nil, fmt.Errorf("azure HTTP %q with malformed error response: %s", resp.Status, err)
+			return nil, errorsource.DownstreamError(fmt.Errorf("azure HTTP %q with malformed error response: %s", resp.Status, err), false)
 		}
-		return nil, fmt.Errorf("azure HTTP %q: %q", resp.Status, r.Error.Message)
+		return nil, errorsource.SourceError(backend.ErrorSourceFromHTTPStatus(resp.StatusCode), fmt.Errorf("azure HTTP %q: %q", resp.Status, r.Error.Message), false)
 	}
 	var clusterData struct {
 		Data []struct {