+++ pkg/azuredx/models/table.go
@@ -12,6 +12,7 @@ import (
 	"time"
 
 	"github.com/grafana/grafana-plugin-sdk-go/data"
+	"github.com/grafana/grafana-plugin-sdk-go/experimental/errorsource"
 	jsoniter "github.com/json-iterator/go"
 )
 
@@ -554,16 +555,16 @@ func TableFromJSON(rc io.Reader) (*TableResponse, error) {
 	if err != nil {
 		return nil, err
 	}
-	if tr.Tables == nil || len(tr.Tables) == 0 {
+	if len(tr.Tables) == 0 {
 		return nil, fmt.Errorf("unable to parse response, parsed response has no tables")
 	}
 
-	if tr.Exceptions != nil && len(tr.Exceptions) > 0 {
+	if len(tr.Exceptions) > 0 {
 		errMsg := ""
 		for _, e := range tr.Exceptions {
 			errMsg += e + ". "
 		}
-		return nil, errors.New(errMsg)
+		return nil, errorsource.DownstreamError(errors.New(errMsg), false)
 	}
 
 	return tr, nil